requirements.txt
django==3.2.9
djangorestframework==3.12.4
drf_yasg==1.20.0
psycopg2-binary==2.9.1
psycopg2
